services/elevenlabs-service/requirements.txt
# ElevenLabs Python SDK and dependencies
elevenlabs>=1.6.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
httpx>=0.25.0

# Audio processing
pydub>=0.25.1
librosa>=0.10.1
soundfile>=0.12.1

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2